run gpt2_normal

+++ Utils/NLP_Models/GPT2.py
@@ -20,10 +20,10 @@ def start(jira_name, train_data_list, test_data_list, model_name, path_to_save):
         test_data = test_data_list[f'{k_unstable}']
 
         x_train = train_data['sentence']
-        y_train = train_data['label'].float()
+        y_train = train_data['label'].astype(float)
 
         x_test = test_data['sentence']
-        y_test = test_data['label'].float()
+        y_test = test_data['label'].astype(float)
 
         x_train_tokenized = tokenizer.batch_encode_plus(
             x_train,
